move chat completion route
